+++ js/data.js
@@ -1,0 +1,44 @@
+const descriptions = [
+  'Рассвет в горах',
+  'Улица старого города',
+  'Кошки на подоконнике',
+  'Осенний парк',
+  'Морской шторм',
+  'Уютное кафе',
+  'Зимний лес',
+  'Город ночью',
+  'Поле с цветами',
+  'Река в тумане',
+  'Закат на пляже',
+  'Кофе и книга',
+  'Горная тропа',
+  'Дождь в городе',
+  'Снежные вершины',
+  'Мост через реку',
+  'Сакура весной',
+  'Пустынный пейзаж',
+  'Водопад в лесу',
+  'Северное сияние'
+];
+
+const commentsList = [
+  'Отличное фото!',
+  'Какой красивый кадр!',
+  'Обожаю такие моменты',
+  'Прелесть просто',
+  'Идеально снято',
+  'Милый котик',
+  'Хочу туда же',
+  'Невероятные цвета',
+  'Это просто волшебство',
+  'Сохранил себе',
+  'Лучшее фото дня',
+  'Как атмосферно!',
+  'Душа отдыхает',
+  'Потрясающе',
+  'Обожаю такие пейзажи'
+];
+
+const names = ['Артём', 'Александра', 'Петр', 'Никита', 'Виктория', 'Савелий', 'Анна', 'Дарья'];
+
+export { descriptions, commentsList, names };
